+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/languages/LanguagePluginsService.kt
@@ -4,6 +4,7 @@ import org.jetbrains.bsp.bazel.info.BspTargetInfo
 import org.jetbrains.bsp.bazel.server.sync.languages.cpp.CppLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.languages.cpp.CppModule
 import org.jetbrains.bsp.bazel.server.sync.languages.java.JavaLanguagePlugin
+import org.jetbrains.bsp.bazel.server.sync.languages.kotlin.KotlinLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.languages.scala.ScalaLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.languages.thrift.ThriftLanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.model.Language
@@ -13,6 +14,7 @@ class LanguagePluginsService(
     val scalaLanguagePlugin: ScalaLanguagePlugin,
     val javaLanguagePlugin: JavaLanguagePlugin,
     val cppLanguagePlugin: CppLanguagePlugin,
+    val kotlinLanguagePlugin: KotlinLanguagePlugin,
     private val thriftLanguagePlugin: ThriftLanguagePlugin
 ) {
     private val emptyLanguagePlugin: EmptyLanguagePlugin = EmptyLanguagePlugin()
@@ -27,7 +29,8 @@ class LanguagePluginsService(
     fun getPlugin(languages: Set<Language>): LanguagePlugin<*> =
         when {
             languages.contains(Language.SCALA) -> scalaLanguagePlugin
-            (languages.contains(Language.JAVA) || languages.contains(Language.KOTLIN)) -> javaLanguagePlugin
+            languages.contains(Language.JAVA) -> javaLanguagePlugin
+            languages.contains(Language.KOTLIN) -> kotlinLanguagePlugin
             languages.contains(Language.CPP) -> cppLanguagePlugin
             languages.contains(Language.THRIFT) -> thriftLanguagePlugin
             else -> emptyLanguagePlugin